# packages/contracts/openapi/jarvis.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JARVIS X",
    "description": "Autonomous personal operations platform. Observes commitments, predicts failure, prepares the next best action, executes through policy-controlled tools, verifies the result, and escalates only when authorized.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/connectors": {
      "get": {
        "tags": [
          "connectors"
        ],
        "summary": "List Connectors",
        "description": "What is connected, what it can see, and how much it has stored.",
        "operationId": "list_connectors_v1_connectors_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Connectors V1 Connectors Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/connectors/google/authorize": {
      "get": {
        "tags": [
          "connectors"
        ],
        "summary": "Google Authorize",
        "description": "Begin the Google flow. Returns the URL to open in a browser.",
        "operationId": "google_authorize_v1_connectors_google_authorize_get",
        "parameters": [
          {
            "name": "include_write",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Also request send/create scopes. Off by default: outbound permission is requested only when a feature that needs it is enabled.",
              "default": false,
              "title": "Include Write"
            },
            "description": "Also request send/create scopes. Off by default: outbound permission is requested only when a feature that needs it is enabled."
          },
          {
            "name": "include_youtube",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Also request youtube.upload, for the video pipeline.",
              "default": false,
              "title": "Include Youtube"
            },
            "description": "Also request youtube.upload, for the video pipeline."
          },
          {
            "name": "include_classroom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Also request Classroom read scopes. Separate because a school account that has not authorized this app fails the whole consent screen.",
              "default": false,
              "title": "Include Classroom"
            },
            "description": "Also request Classroom read scopes. Separate because a school account that has not authorized this app fails the whole consent screen."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Google Authorize V1 Connectors Google Authorize Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/connectors/google/callback": {
      "get": {
        "tags": [
          "connectors"
        ],
        "summary": "Google Callback",
        "description": "Where Google sends the user back.\n\nUnauthenticated by necessity \u2014 the browser arrives here from Google, not from the app \u2014\nso ``state`` is the only proof of who started the flow.",
        "operationId": "google_callback_v1_connectors_google_callback_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "State"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/connectors/sync": {
      "post": {
        "tags": [
          "connectors"
        ],
        "summary": "Sync All",
        "description": "Scan every connected account now, and say what was found.",
        "operationId": "sync_all_v1_connectors_sync_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Sync All V1 Connectors Sync Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/connectors/{account_id}/sync": {
      "post": {
        "tags": [
          "connectors"
        ],
        "summary": "Sync One",
        "operationId": "sync_one_v1_connectors__account_id__sync_post",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Sync One V1 Connectors  Account Id  Sync Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/connectors/{account_id}/disconnect": {
      "post": {
        "tags": [
          "connectors"
        ],
        "summary": "Disconnect",
        "description": "Disconnect, and by default delete what was stored.\n\nDeleting is the default rather than an extra step: a connector that keeps your mail\nafter you disconnect it has not really been disconnected.",
        "operationId": "disconnect_v1_connectors__account_id__disconnect_post",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          },
          {
            "name": "delete_data",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Also delete everything fetched through this connector.",
              "default": true,
              "title": "Delete Data"
            },
            "description": "Also delete everything fetched through this connector."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Disconnect V1 Connectors  Account Id  Disconnect Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}